refactor: improve unit command
